Add redfish log support for AC lost power on
The current plan is that only threshold sensor events
will be logged to the IPMI SEL,
and all other events will be logged to the Redfish Event Log.

The first power on after AC lost will be logged into redfish.

Tested:
Disconnect the AC cable and connect it again, power on the host.
Check the redfish logs about powercontrol:
https://BMCIP/redfish/v1/Systems/system/LogServices/EventLog/Entries

Change-Id: I3df6d75b14bfd42da33f0c983b7680f5185cab37
Signed-off-by: Yong Li <yong.b.li@linux.intel.com>