Add the OperatingSystem.Status interface
The OperatingSystem.Status interface shows the current status of
the OS. Pre-OS it will be "Inactive". When the BIOS has reached
POST complete and is ready to start the OS, it will move to
"Standby".

Tested:
Started power control with POST complete asserted and de-asserted
and confirmed that "OperatingSystemState" is initialized correctly.

Used 'ipmitool power on' and 'ipmitool power off' and confirmed that
"OperatingSystemState" changes to the correct state when POST complete
changes.

Change-Id: I46be0d7aac30e01500ca66c50042c2af7bb919fb
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>