Add SIO S5 power control events
SIO S5 is used to differentiate between a normal OS-initiated
shutdown and a power loss.

This change adds the SIO S5 assert and de-assert events and handles
them as a normal transition to off.

Tested:
Initiated a shutdown from UEFI and confirmed that the SIO S5 signal
asserts and begins the transition to off rather than being treated
as a power loss.

Change-Id: Id497ba0f7c6355252861735314f780f1d04fa7a6
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>