Use the ACBoot dbus property instead of checking PFail directly
This changes power restore policy to start based on the ACBoot
dbus property instead of directly checking the PFail bit.

Tested:
Verified that with power restore policy set to always-on, on a
BMC reboot, the power restore policy is not started and on an AC
cycle, power restore policy starts and boots the host.

Change-Id: I1c0aaa1d1fef2f8a67c6aa2a07151d23b4fd6328
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>