Track power-state instead of power-drop for restore policy
Since there is a race condition to save the power-drop status before
power goes down, switch to tracking power-state using a delay. This
will make sure that if the BMC loses power during a transition that
the old state will remain saved.

For example, during an AC loss when both the host and BMC are losing
power, the delay will cause the BMC to not update the host state to
Off before losing power.

Tested:
Verified that with power restore policy set to previous, removing
AC power while the host is On results in restoring to the On state.

Change-Id: Iec2189d06306badb05cd719f253099a71d583fc1
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>