Handle SIO S5 de-assert events while Off
If the PCH initiates a wake from Off, it will de-assert the SIO
S5 signal. This change makes the BMC accept that event and move to
the Wait for Power Supply Power OK state.

Tested:
Executed 'reset -c' from the UEFI shell and confirmed that the BMC
correctly recognized the event and transitioned to the On state.

Change-Id: Ib82108cfbb5697a26a08030884542a4a061acc78
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>