Add host POH Counter support
The POH shows how many hours the host system has been running.
If the host state is running, it will be incremented by one for
every hour. It won't get updated when the host state is not running.

Tested:
Power on the host and wait for several hours, run the below command:
ipmitool chassis poh
POH Counter  : 0 days, 3 hours

Change-Id: I97a9ea79be4611ccd563dce6544bc18af1acf0a8
Signed-off-by: Yong Li <yong.b.li@linux.intel.com>