Set the restart cause on warm resets from host software
The only signal we currently have to indicate a warm reset from
host software is the POST Complete signal.

This change sets the restart cause when POST Complete de-asserts
only in the "On" state when the reset button has not been pressed.

Tested:
Booted to UEFI shell and pressed CTRL-ALT-DEL and confirmed that
the restart cause changed to 0x0a.
Pressed the reset button and confirmed that the restart cause
changed to 0x02.
Executed 'reset -w' at the UEFI shell and confirmed that the
restart cause changed to 0x0a.

Change-Id: I16ca3a72b6279e14ffb31d441b7236fdec4847f5
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>